+++ src/opender/__init__.py
@@ -1,4 +1,4 @@
-__version__ = '1.0.0'
+__version__ = '1.0.1'
 
 from .der import DER
 from .common_file_format import DERCommonFileFormat