+++ src/opender/auxiliary_funcs/low_pass_filter.py
@@ -12,10 +12,9 @@
 #   prior written permission.
 
 
+from opender import der
 
-# -*- coding: utf-8 -*-
 
-from . import der
 class LowPassFilter:
     """
     |  Low pass filter for modeling DER open loop response behavior and other related responses
@@ -40,10 +39,10 @@ class LowPassFilter:
         
         :param lpf_out:   Low pass filtered result
         """
-        if self.lpf_in_prev == None:
+        if self.lpf_in_prev is None:
             self.lpf_in_prev = lpf_in
 
-        if self.lpf_out_prev == None:
+        if self.lpf_out_prev is None:
             self.lpf_out_prev = lpf_in
 
         if(t_olrt < (1.15 * der.DER.t_s)):
@@ -53,6 +52,7 @@ class LowPassFilter:
             t_olrt_t = t_olrt/1.15
             lpf_out = ((der.DER.t_s / (der.DER.t_s + t_olrt_t)) * (lpf_in + self.lpf_in_prev)) + ((t_olrt_t - der.DER.t_s) / (
                         der.DER.t_s + t_olrt_t) * self.lpf_out_prev)
+
         self.lpf_in_prev = lpf_in
         self.lpf_out_prev = lpf_out
         return lpf_out