+++ src/opender/auxiliary_funcs/ramping.py
@@ -12,13 +12,11 @@
 #   prior written permission.
 
 
-# -*- coding: utf-8 -*-
 """
-
 @author: Jithendar Anandan
 @email: janandan@epri.com
 """
-from . import der
+from opender import der
 class Ramping:
     """
     |  Ramp rate limit function