+++ src/opender/reactive_power_support_funcs/constant_power_factor.py
@@ -16,7 +16,7 @@
 # -*- coding: utf-8 -*-
 
 import math
-from .low_pass_filter import LowPassFilter
+from opender.auxiliary_funcs.low_pass_filter import LowPassFilter
 
 
 class ConstantPowerFactor:
@@ -26,14 +26,14 @@ class ConstantPowerFactor:
     """
 
     def __init__(self):
-        self.pf_olrt = LowPassFilter()
+        self.pf_lpf = LowPassFilter()
 
     def calculate_q_const_pf_desired_kvar(self,der_file, exec_delay, p_desired_kw):
         """
         Calculates and returns output reactive power from Constant Power Factor function
 
         Variable used in this function:
-        
+
         :param p_desired_kw:  Desired output active power considering DER enter service performance
         :param const_pf_exec:  Constant Power Factor Setting (CONST_PF) after execution delay
         :param const_pf_excitation_exec:  Constant Power Factor Excitation (CONST_PF_EXCITATION) after execution delay
@@ -48,19 +48,20 @@ class ConstantPowerFactor:
         :param q_const_pf_desired_kvar:	Output reactive power from constant power factor function
         """
 
-        #Eq. 35, calculate reactive power reference according to desired active power and constant power factor setting
-        if(exec_delay.const_pf_excitation_exec == "INJ"):
-            q_const_pf_desired_ref_kvar = p_desired_kw * (math.sqrt(1 - (exec_delay.const_pf_exec * exec_delay.const_pf_exec))/exec_delay.const_pf_exec)
+        # Eq. 35, calculate reactive power reference according to desired active power and constant power factor setting
+        if exec_delay.const_pf_excitation_exec == "INJ":
+            q_const_pf_desired_ref_kvar = p_desired_kw * \
+                                          (math.sqrt(1 - (exec_delay.const_pf_exec ** 2))/exec_delay.const_pf_exec)
+        elif exec_delay.const_pf_excitation_exec == "ABS":
+            q_const_pf_desired_ref_kvar = -p_desired_kw * \
+                                          (math.sqrt(1 - (exec_delay.const_pf_exec ** 2))/exec_delay.const_pf_exec)
+        else:
+            print(f'CONST_PF_EXCITATION value unexpected:{exec_delay.const_pf_excitation_exec}')
+            q_const_pf_desired_ref_kvar = 0
 
-        if(exec_delay.const_pf_excitation_exec == "ABS"):
-            q_const_pf_desired_ref_kvar = p_desired_kw * (math.sqrt(1 - (exec_delay.const_pf_exec * exec_delay.const_pf_exec))/exec_delay.const_pf_exec)
-            q_const_pf_desired_ref_kvar = -q_const_pf_desired_ref_kvar
-
-        '''
-        Eq. 36, apply the low pass filter to the reference reactive power. Note that there can be multiple different 
-        ways to implement this behavior in an actual DER. The model may be updated in a future version, according to the
-        lab test results.
-        '''
-        q_const_pf_desired_kvar = self.pf_olrt.low_pass_filter(q_const_pf_desired_ref_kvar, der_file.CONST_PF_RT)
+        # Eq. 36, apply the low pass filter to the reference reactive power. Note that there can be multiple different
+        # ways to implement this behavior in an actual DER. The model may be updated in a future version, according to
+        # the lab test results.
+        q_const_pf_desired_kvar = self.pf_lpf.low_pass_filter(q_const_pf_desired_ref_kvar, der_file.CONST_PF_RT)
 
         return q_const_pf_desired_kvar