+++ CHANGELOG.rst
@@ -1,6 +1,13 @@
 
 Changelog
 =========
+Unreleased updates
+-----------------
+* Improve model input validity check process
+
+1.0.1 (2022-06-17)
+-----------------
+* Include parameter csv files into the PyPI release
 
 1.0.0 (2022-05-17)
 ------------------